+++ 08.ls_object/lib/ls_long_format.rb
@@ -1,22 +1,6 @@
 # frozen_string_literal: true
 
 require 'etc'
-require 'pathname'
-
-class LsCommandExecution
-  def initialize(filenames, long_format: false, reverse: false, dot_match: false)
-    @filenames = filenames
-    @long_format = long_format
-    @reverse = reverse
-    @dot_match = dot_match
-  end
-
-  def run_ls
-    @filenames = @dot_match ? Dir.glob('*', File::FNM_DOTMATCH).sort : Dir.glob('*').sort
-    @filenames = @reverse ? @filenames.reverse : @filenames
-    @long_format ? LsLongFormat.new(@filenames).ls_long_format_display : LsShortFormat.new(@filenames).ls_short_format_display
-  end
-end
 
 FILE_TYPES = {
   file: '-',
@@ -51,6 +35,8 @@ class LsLongFormat < LsCommandExecution
     body = render_long_format_body(row_data)
     [total, *body].join("\n")
   end
+
+  private
 
   def build_data(filename, stat)
     {
@@ -128,35 +114,5 @@ class LsLongFormat < LsCommandExecution
 
   def mtime(file_stat)
     file_stat.mtime.strftime("  %-m %e %H:%M ")
-  end
-end
-
-WIDTH = 80 # widthを80に固定
-MAX_FILENAME_COUNT = 21 # 横幅の文字最大21文字に設定
-
-class LsShortFormat < LsCommandExecution
-
-  def ls_short_format_display
-    transposed_file_names = safe_transpose(@filenames.each_slice(row_count).to_a)
-    format_table(transposed_file_names)
-  end
-
-  def columns
-    WIDTH / MAX_FILENAME_COUNT
-  end
-
-  def row_count
-    (@filenames.count.to_f / columns).ceil
-  end
-
-  def safe_transpose(filenames)
-    first, *rest = filenames
-    first.zip(*rest)
-  end
-
-  def format_table(filenames)
-    filenames.map do |row_files|
-      row_files.map { |filename| filename.to_s.ljust(MAX_FILENAME_COUNT) }.join.rstrip
-    end.join("\n")
   end
 end
